+++ src/lib.rs
@@ -104,7 +104,8 @@ pub mod types {
     };
     pub use super::emails::types::{
         Attachment, CancelScheduleResponse, ContentOrPath, CreateEmailBaseOptions,
-        CreateEmailResponse, Email, EmailId, Tag, UpdateEmailOptions, UpdateEmailResponse,
+        CreateEmailResponse, Email, EmailEvent, EmailId, Tag, UpdateEmailOptions,
+        UpdateEmailResponse,
     };
     pub use super::error::types::{ErrorKind, ErrorResponse};
 }